Wed, Aug 3, 2022, 10:34 PM +00:00

+++ psets/pset4/filter-less/helpers.c
@@ -17,12 +17,12 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
         for (int j = 0; j < width; j++)
         {
             // Get all 3 rgb colors
-            double red = image[i][j].rgbtRed;
-            double green = image[i][j].rgbtGreen;
-            double blue = image[i][j].rgbtBlue;
+            int red = image[i][j].rgbtRed;
+            int green = image[i][j].rgbtGreen;
+            int blue = image[i][j].rgbtBlue;
 
             // Calculate the average and round
-            int average = round((red + green + blue) / 3);
+            double average = round((red + green + blue) / 3);
 
             // If the average is higher than 255 set to 255
             if (average > 255)
